+++ EdeUNLa/src/datos/ItemFactura.java
@@ -2,7 +2,7 @@ package datos;
 
 public class ItemFactura {
 
-	private int idItemFactura;
+	private long idItemFactura;
 	private String detalle;
 	private double precioUnitario;
 	private int cantidad;
@@ -27,11 +27,11 @@ public class ItemFactura {
 		this.lectura = lectura;
 	}
 
-	public int getIdItemFactura() {
+	public long getIdItemFactura() {
 		return idItemFactura;
 	}
 
-	protected void setIdItemFactura(int idItemFactura) {
+	protected void setIdItemFactura(long idItemFactura) {
 		this.idItemFactura = idItemFactura;
 	}
